Extract UI actions into `PatientEntryUiActions` from `PatientEntryUi`
