+++ app/src/main/java/org/simple/clinic/newentry/PatientEntryUiActions.kt
@@ -1,3 +1,13 @@
 package org.simple.clinic.newentry
 
-interface PatientEntryUiActions
+import org.simple.clinic.newentry.country.InputFields
+import org.simple.clinic.patient.OngoingNewPatientEntry
+
+interface PatientEntryUiActions {
+  fun prefillFields(entry: OngoingNewPatientEntry)
+  fun scrollToFirstFieldWithError()
+  fun scrollFormOnGenderSelection()
+  fun setShowDatePatternInDateOfBirthLabel(showPattern: Boolean)
+  fun openMedicalHistoryEntryScreen()
+  fun setupUi(inputFields: InputFields)
+}
